move puzzle specific code to its module
