+++ 07/amplification_circuit/lib/amplification_circuit.ex
@@ -2,17 +2,28 @@ defmodule AmplificationCircuit do
   @moduledoc """
   Documentation for AmplificationCircuit.
   """
+  defmodule Lista do
+    def permutations([]), do: [[]]
 
-  @doc """
-  Hello world.
+    def permutations(list) do
+      for(elem <- list, rest <- permutations(list -- [elem]), do: [elem | rest])
+    end
+  end
 
-  ## Examples
+  def find_highest_amplifiers_output_signal(phase_settings_list, program) do
+    Enum.map(Lista.permutations(phase_settings_list), fn phase_sequence ->
+      get_amplifiers_output(phase_sequence, program)
+    end)
+    |> Enum.max()
+  end
 
-      iex> AmplificationCircuit.hello()
-      :world
+  def get_amplifiers_output(phase_sequence, program) do
+    {_, exit_code} =
+      Enum.map_reduce(phase_sequence, 0, fn x, exit_code ->
+        {_result, _program, exit_code} = IntcodeComputer.run(program, [x, exit_code])
+        {x, exit_code}
+      end)
 
-  """
-  def hello do
-    :world
+    exit_code
   end
 end
